kata-deploy: Configure containerd atomically to avoid corruption
we might be not the only ones modifying containerd configuration or
kata-deploy can be re-started (especially since we are restarting
various services) leading to corrupted or partially modified
configuration. Let's create a tmp file and work on it instead, replacing
the primary configuration when we are done.

For cri-o we likely don't need this since we don't modify the primary
config file and only add drop-in files, which gets always recreated from
scratch.

Signed-off-by: Lukáš Doktor <ldoktor@redhat.com>